[FIX] 회원탈퇴: 조건 설정, 일반유저-드라이버 로직 분리

+++ src/main/java/mallang_trip/backend/domain/user/service/UserWithdrawalService.java
@@ -1,11 +1,20 @@
 package mallang_trip.backend.domain.user.service;
 
 
+import static mallang_trip.backend.domain.party.constant.DriverPenaltyStatus.PENALTY_EXISTS;
 import static mallang_trip.backend.domain.party.exception.PartyExceptionStatus.ONGOING_PARTY_EXISTS;
+import static mallang_trip.backend.domain.user.constant.Role.ROLE_DRIVER;
+import static mallang_trip.backend.domain.user.constant.Role.ROLE_USER;
+import static mallang_trip.backend.domain.user.exception.UserExceptionStatus.PAYMENT_FAILED_EXISTS;
+import static mallang_trip.backend.global.io.BaseResponseStatus.Bad_Request;
 
 import lombok.RequiredArgsConstructor;
+import mallang_trip.backend.domain.driver.constant.DriverStatus;
+import mallang_trip.backend.domain.driver.entity.Driver;
+import mallang_trip.backend.domain.driver.service.DriverService;
 import mallang_trip.backend.domain.party.constant.ProposalStatus;
 import mallang_trip.backend.domain.party.constant.ProposalType;
+import mallang_trip.backend.domain.reservation.repository.ReservationRepository;
 import mallang_trip.backend.global.io.BaseException;
 import mallang_trip.backend.domain.user.entity.User;
 import mallang_trip.backend.domain.party.repository.PartyProposalRepository;
@@ -22,7 +31,9 @@ public class UserWithdrawalService {
 	private final CurrentUserService currentUserService;
 	private final ChatService chatService;
 	private final PartyRepository partyRepository;
+	private final DriverService driverService;
 	private final PartyProposalRepository partyProposalRepository;
+	private final ReservationRepository reservationRepository;
 
 	public static final String UNKNOWN_USER = "알 수 없음";
 	public static final String EMPTY_STRING = "";
@@ -32,14 +43,47 @@ public class UserWithdrawalService {
 	 */
 	public void withdrawal() {
 		User user = currentUserService.getCurrentUser();
+		if (user.getRole().equals(ROLE_USER)) {
+			withdrawalByUser(user);
+		} else if (user.getRole().equals(ROLE_DRIVER)) {
+			withdrawalByDriver();
+		} else {
+			throw new BaseException(Bad_Request);
+		}
+	}
+
+	/**
+	 * 일반 유저 회원탈퇴
+	 */
+	public void withdrawalByUser(User user) {
+		// 진행중인 파티가 존재할 경우
 		if (isOngoingPartyExists(user)) {
 			throw new BaseException(ONGOING_PARTY_EXISTS);
 		}
-		// TODO: 대금 지급이 이루어지지 않았을 경우 exception handling
-		// TODO: 결제 정보 저장
+		// 미납금이 존재할 경우
+		if (reservationRepository.isPaymentFailedExistsByUser(user.getId())) {
+			throw new BaseException(PAYMENT_FAILED_EXISTS);
+		}
 		deletePersonalInformation(user);
 		chatService.leaveAllChat(user);
-		user.setRefreshToken(null);
+	}
+
+	/**
+	 * 드라이버 회원탈퇴
+	 */
+	public void withdrawalByDriver() {
+		Driver driver = driverService.getCurrentDriver();
+		// 진행중인 파티가 존재할 경우
+		if (partyRepository.isOngoingPartyExistsByDriver(driver.getId())) {
+			throw new BaseException(ONGOING_PARTY_EXISTS);
+		}
+		// 미납 위약금이 존재할 경우
+		if (partyRepository.existsByDriverAndDriverPenaltyStatus(driver, PENALTY_EXISTS)) {
+			throw new BaseException(PAYMENT_FAILED_EXISTS);
+		}
+		driver.changeStatus(DriverStatus.CANCELED);
+		deletePersonalInformation(driver.getUser());
+		chatService.leaveAllChat(driver.getUser());
 	}
 
 	/**
@@ -47,21 +91,25 @@ public class UserWithdrawalService {
 	 */
 	private boolean isOngoingPartyExists(User user) {
 		return partyRepository.isOngoingPartyExists(user.getId())
-				|| partyProposalRepository.existsByProposerAndTypeAndStatus(user,
-				ProposalType.JOIN_WITH_COURSE_CHANGE, ProposalStatus.WAITING);
+			|| partyProposalRepository.existsByProposerAndTypeAndStatus(user,
+			ProposalType.JOIN_WITH_COURSE_CHANGE, ProposalStatus.WAITING);
 	}
 
-	private void deletePersonalInformation(User user){
+	/**
+	 * 개인정보 삭제
+	 */
+	private void deletePersonalInformation(User user) {
 		//user.setBirthDay(null);
 		user.setEmail(EMPTY_STRING);
 		user.setIntroduction(null);
 		user.setLoginId(EMPTY_STRING);
-		user.setName(UNKNOWN_USER);
+		//user.setName(UNKNOWN_USER);
 		user.setNickname(UNKNOWN_USER);
 		user.setPassword(EMPTY_STRING);
 		user.setPhoneNumber(EMPTY_STRING);
 		user.setProfileImage(null);
 		user.setRefreshToken(null);
+		user.setDi(null);
 		user.setDeleted(true);
 	}
